Moved beans to SpringMongoDBConfig
