+++ itachallenge-challenge/src/main/java/com/itachallenge/challenge/config/SpringMongoDBConfig.java
@@ -1,6 +1,16 @@
 package com.itachallenge.challenge.config;
 
+import com.mongodb.ConnectionString;
+import com.mongodb.MongoClientSettings;
+import com.mongodb.reactivestreams.client.MongoClient;
+import com.mongodb.reactivestreams.client.MongoClients;
+import io.mongock.driver.api.driver.ConnectionDriver;
+import io.mongock.driver.mongodb.reactive.driver.MongoReactiveDriver;
+import org.bson.UuidRepresentation;
+import org.bson.codecs.configuration.CodecRegistry;
+import org.bson.codecs.pojo.PojoCodecProvider;
 import org.springframework.beans.factory.BeanFactory;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.mongodb.MongoDatabaseFactory;
@@ -10,8 +20,13 @@ import org.springframework.data.mongodb.core.convert.DefaultMongoTypeMapper;
 import org.springframework.data.mongodb.core.convert.MappingMongoConverter;
 import org.springframework.data.mongodb.core.mapping.MongoMappingContext;
 
+import static org.bson.codecs.configuration.CodecRegistries.fromProviders;
+import static org.bson.codecs.configuration.CodecRegistries.fromRegistries;
+
 @Configuration
 public class SpringMongoDBConfig {
+    @Value("${spring.data.mongodb.uri}")
+    private String mongoConnectionString;
 
     //to avoid _class attribute in mongoDB
     @Bean
@@ -22,6 +37,26 @@ public class SpringMongoDBConfig {
         MappingMongoConverter mappingConverter = new MappingMongoConverter(dbRefResolver, context);
         mappingConverter.setTypeMapper(new DefaultMongoTypeMapper(null));
         return mappingConverter;
+    }
+
+    //Todo to check alternatives
+    @Bean
+    public ConnectionDriver mongockConnection(MongoClient mongoClient) {
+        return MongoReactiveDriver.withDefaultLock(mongoClient, "challenges");
+    }
+
+    //Todo to check alternatives
+    @Bean
+    MongoClient mongoClient() {
+        CodecRegistry codecRegistry = fromRegistries(
+                MongoClientSettings.getDefaultCodecRegistry(),
+                fromProviders(PojoCodecProvider.builder().automatic(true).build()));
+
+        return MongoClients.create(MongoClientSettings.builder()
+                .applyConnectionString(new ConnectionString(mongoConnectionString))
+                .codecRegistry(codecRegistry)
+                .uuidRepresentation(UuidRepresentation.STANDARD)
+                .build());
     }
 
 }
